+++ Makefile
@@ -22,6 +22,7 @@ LEXER_FILES = $(LEXER_DIR)/tokenizer.c \
 			  $(LEXER_DIR)/token/token_utils.c \
 			  $(LEXER_DIR)/handler/expansion_braced_handler.c \
 			  $(LEXER_DIR)/handler/expansion_handler.c \
+			  $(LEXER_DIR)/handler/expansion_param_handler.c \
 			  $(LEXER_DIR)/handler/operator_handler.c \
 			  $(LEXER_DIR)/handler/quote_handler.c \
 			  $(LEXER_DIR)/handler/quote_content_handler.c \
@@ -29,6 +30,7 @@ LEXER_FILES = $(LEXER_DIR)/tokenizer.c \
 			  $(LEXER_DIR)/handler/quote_validator.c \
 			  $(LEXER_DIR)/handler/var_expansion_handler.c \
 			  $(LEXER_DIR)/handler/wildcard_handler.c \
+			  $(LEXER_DIR)/handler/word_extract_handler.c \
 			  $(LEXER_DIR)/handler/word_handler.c \
 			  $(LEXER_DIR)/utils/char_check.c \
 			  $(LEXER_DIR)/utils/expansion_utils.c \
