+++ src/lexer/handler/word_handler.c
@@ -6,7 +6,7 @@
 /*   By: spyun <spyun@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/01/20 15:32:09 by spyun         #+#    #+#                 */
-/*   Updated: 2025/02/21 09:40:12 by spyun         ########   odam.nl         */
+/*   Updated: 2025/02/21 09:47:29 by spyun         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,56 +18,6 @@ static void	skip_spaces(t_tokenizer *tokenizer)
 	while (tokenizer->input[tokenizer->position]
 		&& ft_isspace(tokenizer->input[tokenizer->position]))
 		tokenizer->position++;
-}
-
-/* Expand exit status */
-static char	*get_next_word_part(t_tokenizer *tokenizer)
-{
-	char	*part;
-
-	if (is_quote(tokenizer->input[tokenizer->position]))
-	{
-		part = extract_quoted_content_with_expansion(tokenizer,
-				tokenizer->input[tokenizer->position]);
-	}
-	else
-	{
-		part = ft_substr(tokenizer->input, tokenizer->position, 1);
-		tokenizer->position++;
-	}
-	return (part);
-}
-
-/* Extract word */
-static char	*extract_word(t_tokenizer *tokenizer)
-{
-	char	*result;
-	char	*part;
-	char	*temp;
-
-	result = ft_strdup("");
-	if (!result)
-		return (NULL);
-	while (tokenizer->input[tokenizer->position]
-		&& !ft_isspace(tokenizer->input[tokenizer->position])
-		&& !is_operator(&tokenizer->input[tokenizer->position]))
-	{
-		if (is_exit_status_var(&tokenizer->input[tokenizer->position])
-			&& !tokenizer->in_quote)
-		{
-			part = expand_exit_status();
-			tokenizer->position += 2;
-		}
-		else
-			part = get_next_word_part(tokenizer);
-		if (!part)
-			return (free(result), NULL);
-		temp = join_words(result, part);
-		if (!temp)
-			return (free(result), NULL);
-		result = temp;
-	}
-	return (result);
 }
 
 /* Analyze and create token */
